libft Part 1

+++ libft/ft_strlcat.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
 	unsigned int len_dest;
 	unsigned int len_src;
